+++ packages/cli/src/commands/source/push/add.ts
@@ -1,6 +1,7 @@
-import {Source} from '@coveo/push-api-client';
+import {DocumentBuilder, Source} from '@coveo/push-api-client';
 import {Command, flags} from '@oclif/command';
 import {blueBright, green} from 'chalk';
+import {cli} from 'cli-ux';
 import {readdirSync} from 'fs';
 import path from 'path';
 import dedent from 'ts-dedent';
@@ -14,6 +15,11 @@ import {
 } from '../../../lib/decorators/preconditions';
 import {AuthenticatedClient} from '../../../lib/platform/authenticatedClient';
 import {parseAndGetDocumentBuilderFromJSONDocument} from '../../../lib/push/parseFile';
+import {
+  ErrorFromAPI,
+  errorMessage,
+  successMessage,
+} from '../../../lib/push/userFeedback';
 
 interface AxiosResponse {
   status: number;
@@ -21,7 +27,9 @@ interface AxiosResponse {
 }
 
 export default class SourcePushAdd extends Command {
-  public static description = 'Push a JSON document into a Coveo push source.';
+  private static maxContentLength = 5 * 1024 * 1024;
+  public static description =
+    'Push a JSON document into a Coveo Push source. See https://github.com/coveo/cli/wiki/Pushing-JSON-files-with-Coveo-CLI for more information.';
 
   public static flags = {
     file: flags.string({
@@ -55,9 +63,10 @@ export default class SourcePushAdd extends Command {
     const {flags} = this.parse(SourcePushAdd);
     if (!flags.file && !flags.folder) {
       this.error(
-        'You must provide either --file= or --folder=. See source:push:add --help for more information.'
+        'You must minimally set the `file` or the `folder` flag. Use `source:push:add --help` to get more information.'
       );
     }
+
     const cfg = await new AuthenticatedClient().cfg.get();
     const source = new Source(cfg.accessToken!, cfg.organization);
 
@@ -82,37 +91,53 @@ export default class SourcePushAdd extends Command {
       return files.map((f) => `${path.join(folder, f)}`);
     });
 
-    const docBuilders = flags.folder.flatMap((folder) => {
-      return readdirSync(folder).flatMap((file) => {
-        const fullPath = path.join(folder, file);
-        const docBuilders =
-          parseAndGetDocumentBuilderFromJSONDocument(fullPath);
-        this.successMessageOnParseFile(fullPath, docBuilders.length);
-        return docBuilders;
-      });
-    });
+    const {send, close} = this.splitByChunkAndUpload(
+      source,
+      args.sourceId,
+      fileNames
+    );
 
-    const res = await source.batchUpdateDocuments(args.sourceId, {
-      addOrUpdate: docBuilders,
-      delete: [],
-    });
-    this.successMessageOnAdd(fileNames, docBuilders.length, res);
+    cli.action.start('Processing...');
+
+    await Promise.all(
+      flags.folder.flatMap((folder) => {
+        return Promise.all(
+          readdirSync(folder).flatMap(async (file) => {
+            const fullPath = path.join(folder, file);
+            const docBuilders =
+              parseAndGetDocumentBuilderFromJSONDocument(fullPath);
+            this.successMessageOnParseFile(fullPath, docBuilders.length);
+            await send(docBuilders);
+          })
+        );
+      })
+    );
+
+    await close();
+    cli.action.stop();
   }
 
   private async doPushFile(source: Source) {
     const {flags, args} = this.parse(SourcePushAdd);
 
-    const docBuilders = flags.file.flatMap((file) => {
-      const docBuilders = parseAndGetDocumentBuilderFromJSONDocument(file);
-      this.successMessageOnParseFile(file, docBuilders.length);
-      return docBuilders;
-    });
+    const {send, close} = this.splitByChunkAndUpload(
+      source,
+      args.sourceId,
+      flags.file
+    );
 
-    const res = await source.batchUpdateDocuments(args.sourceId, {
-      addOrUpdate: docBuilders,
-      delete: [],
-    });
-    this.successMessageOnAdd(flags.file, docBuilders.length, res);
+    cli.action.start('Processing...');
+
+    await Promise.all(
+      flags.file.flatMap(async (file) => {
+        const docBuilders = parseAndGetDocumentBuilderFromJSONDocument(file);
+        this.successMessageOnParseFile(file, docBuilders.length);
+        await send(docBuilders);
+      })
+    );
+
+    await close();
+    cli.action.stop();
   }
 
   private successMessageOnAdd(
@@ -128,14 +153,17 @@ export default class SourcePushAdd extends Command {
       fileNames += ` and ${files.length - 5} more ...`;
     }
 
-    this.log(
-      dedent(`
-    Success: ${green(numAdded)} documents accepted by the Push API from ${green(
-        fileNames
-      )}
-    Status code: ${green(res.status, res.statusText)}
-    `)
+    return successMessage(
+      this,
+      `Success: ${green(
+        numAdded
+      )} documents accepted by the Push API from ${green(fileNames)}.`,
+      res
     );
+  }
+
+  private errorMessageOnAdd(e: ErrorFromAPI) {
+    return errorMessage(this, 'Error while trying to add document.', e);
   }
 
   private successMessageOnParseFile(file: string, numParsed: number) {
@@ -153,5 +181,62 @@ export default class SourcePushAdd extends Command {
       buildAnalyticsFailureHook(this, flags, err)
     );
     throw err;
+  }
+
+  private splitByChunkAndUpload(
+    source: Source,
+    sourceId: string,
+    fileNames: string[],
+    accumulator = this.accumulator
+  ) {
+    const send = async (documentBuilders: DocumentBuilder[]) => {
+      for (const docBuilder of documentBuilders) {
+        const sizeOfDoc = Buffer.byteLength(
+          JSON.stringify(docBuilder.marshal())
+        );
+
+        if (accumulator.size + sizeOfDoc >= SourcePushAdd.maxContentLength) {
+          await this.uploadBatch(
+            source,
+            sourceId,
+            accumulator.chunks,
+            fileNames
+          );
+          accumulator.chunks = [docBuilder];
+          accumulator.size = sizeOfDoc;
+        } else {
+          accumulator.size += sizeOfDoc;
+          accumulator.chunks.push(docBuilder);
+        }
+      }
+    };
+    const close = async () => {
+      await this.uploadBatch(source, sourceId, accumulator.chunks, fileNames);
+    };
+    return {send, close};
+  }
+
+  private async uploadBatch(
+    source: Source,
+    sourceId: string,
+    batch: DocumentBuilder[],
+    fileNames: string[]
+  ) {
+    try {
+      const res = await source.batchUpdateDocuments(sourceId, {
+        addOrUpdate: batch,
+        delete: [],
+      });
+      this.successMessageOnAdd(fileNames, batch.length, res);
+    } catch (e) {
+      this.errorMessageOnAdd(e);
+    }
+  }
+
+  private get accumulator(): {size: number; chunks: DocumentBuilder[]} {
+    return {
+      size: 0,
+      chunks: [],
+    };
   }
 }